refactor: move classes to specific files
