+++ backend/individual-assignment/typescript/src/address.ts
@@ -1,0 +1,1 @@
+export class Address {}
