+++ backend/individual-assignment/typescript/src/customer.ts
@@ -1,0 +1,1 @@
+export class Customer {}
